Match games achievement provider on Plugin/emulator instead of Source field
Also significant refactor of achievement provider config validation
